+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version of inference and switching logic taken from internal Emotech
 code
 - Added in `end_silence_length` to track the raw end silences
+- Added new pub function `validate_input` for `VadSession` struct. `process` function will use it to make sure input is valid in debug mode.